+++ tests/testJIT.cpp
@@ -44,7 +44,7 @@ int main(int argc, const char **argv)
 
     xcc::Parser parser(SM, ctx, ig);
 
-    unsigned num_typedefs, num_tags;
+    size_t num_typedefs, num_tags;
     auto ast = parser.run(num_typedefs, num_tags);
     if (ctx.printer->NumErrors)
         return 1;
