+++ tests/testParser.cpp
@@ -1,5 +1,5 @@
 #include "common.h"
-    unsigned num_typedefs, num_tags;
+    size_t num_typedefs, num_tags;
     auto ast = parser.run(num_typedefs, num_tags);
     if (ctx.printer->NumErrors)
         return 1;
